+++ src/Types/AbstractList.php
@@ -58,4 +58,24 @@ abstract class AbstractList implements Type
     {
         return $this->valueType;
     }
+
+    /**
+     * Returns a rendered output of the Type as it would be used in a DocBlock.
+     */
+    public function __toString() : string
+    {
+        if ($this->keyType) {
+            return 'array<' . $this->keyType . ',' . $this->valueType . '>';
+        }
+
+        if ($this->valueType instanceof Mixed_) {
+            return 'array';
+        }
+
+        if ($this->valueType instanceof Compound) {
+            return '(' . $this->valueType . ')[]';
+        }
+
+        return $this->valueType . '[]';
+    }
 }
