+++ src/Types/Array_.php
@@ -24,23 +24,4 @@ namespace phpDocumentor\Reflection\Types;
  */
 final class Array_ extends AbstractList
 {
-    /**
-     * Returns a rendered output of the Type as it would be used in a DocBlock.
-     */
-    public function __toString() : string
-    {
-        if ($this->keyType) {
-            return 'array<' . $this->keyType . ',' . $this->valueType . '>';
-        }
-
-        if ($this->valueType instanceof Mixed_) {
-            return 'array';
-        }
-
-        if ($this->valueType instanceof Compound) {
-            return '(' . $this->valueType . ')[]';
-        }
-
-        return $this->valueType . '[]';
-    }
 }
