+++ src/Types/Expression_.php
@@ -13,12 +13,33 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Reflection\Types;
 
+use phpDocumentor\Reflection\Type;
+
 /**
- * Represents an array type as described in the PSR-5, the PHPDoc Standard.
+ * Represents an expression type as described in the PSR-5, the PHPDoc Standard.
  *
  */
-final class Expression_ extends AbstractList
+final class Expression_ implements Type
 {
+    /** @var Type */
+    protected $valueType;
+
+    /**
+     * Initializes this representation of an array with the given Type.
+     */
+    public function __construct(Type $valueType)
+    {
+        $this->valueType      = $valueType;
+    }
+
+    /**
+     * Returns the value for the keys of this array.
+     */
+    public function getValueType() : Type
+    {
+        return $this->valueType;
+    }
+
     /**
      * Returns a rendered output of the Type as it would be used in a DocBlock.
      */
